pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "job-automation"
version = "0.1.0"
description = "Minimal Glints scraper using browser-use and a local OpenCode ACP model"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "browser-use",
  "httpx",
  "pydantic>=2",
  "python-dotenv",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
